Demo moving hook into a function
