+++ packages/cli-tool/src/commands/generate/questions.ts
@@ -1,0 +1,21 @@
+import getChoices from '../../helpers/choices'
+import {UI_FRAMEWORK_OPTIONS} from './ui-framework/index'
+import {VERSION_CONTROL_OPTIONS} from './version-control/index'
+
+const uiFrameworkChoices = getChoices(UI_FRAMEWORK_OPTIONS)
+const versionControlChoices = getChoices(VERSION_CONTROL_OPTIONS)
+
+export const questions = [
+  {
+    type: 'list',
+    name: 'versionControl',
+    message: 'Select a version control service:',
+    choices: versionControlChoices,
+  },
+  {
+    type: 'list',
+    name: 'uiFramework',
+    message: 'Select a UI Framework:',
+    choices: uiFrameworkChoices,
+  },
+]
