+++ packages/cli-tool/src/commands/generate/ui-framework/index.ts
@@ -1,0 +1,5 @@
+export const UI_FRAMEWORK_OPTIONS: { [key: string]: string } = {
+  bootstrap: 'Bootstrap',
+  tailwind: 'Tailwind CSS',
+  none: 'None',
+}
