+++ packages/cli-tool/src/commands/generate/version-control/index.ts
@@ -1,0 +1,5 @@
+export const VERSION_CONTROL_OPTIONS: { [key: string]: string } = {
+  github: 'GitHub',
+  gitlab: 'GitLab',
+  none: 'None',
+}
